Make co-reloading more invisible
The previous solution relying on `require` and the fact that a CLJS
watcher already sort by dependency order was somewhat fine but left a
margin of error user-wise.

Ultimately, the new solution leverages the full extent of
`clojure.tools.namespace.*` (unlike previous quick attempts relying only
on the `repl` namespace by coupling it with a file watcher.