Start implementing a plugin scheme for co-loading
It is hard to coordinate different tools relying on co-loading unless
they are centralized in one place. Instead of tools proposing their own
hooks to users, it is best to configure them as plugins in the Medium
co-loading hooks.

Furthermore, it allows for passing information to plugins such as what
namespaces will be reloaded.